01/11/2018 --- Bug fixe. - Create Spawn Robot function in RobotWarsGameMode_BP

+++ RobotWars/Source/RobotWars/Public/EnergySystem.h
@@ -50,7 +50,7 @@ public:
 	///********** VARIABLE **************************************************************************************************************************************************
 private:
 	float CurrentEnergy[NUM_ENERGY_SYSTEMS];
-	int32 CurrentGeneratorStructure;
+	float CurrentGeneratorStructure;
 	int32 EnergyChargeRatePerSecond[NUM_ENERGY_SYSTEMS];
 
 	TArray<SYSTEM> SystemPriority;
